기존의 템플릿 정상여부 체크를 핸드쉐이크로 변경
